// src/app/[locale]/(platform)/fundamentals/html5/basic-elements-and-tags/_components/BasicElementsAndTags/types.ts
export interface BasicElementsAndTagsTypes {
  data: Data;
}

interface Data {
  title: string;
  description: string;
  items: Item2[];
  conclusion: string;
}

interface Item2 {
  title: string;
  description: string;
  items?: Item[];
}

interface Item {
  title: string;
  description: string;
}